Refactor helper functions and update docs
